feat(Rocket): small refactor
